Update project to use TypeScript instead of JS
This commit includes changes to convert JavaScript files into TypeScript and update configurations as necessary. This includes modifications to the scripts, and renaming of several files into .ts extensions. A new ESLint configuration file is added to replace the previous one, providing support for TypeScript. Packages were updated to use TypeScript as well, reducing any potential incompatibility issues. This was done to enhance the type-safety and maintainability of the code. Application functionalities remain unchanged.