Update build scripts and refactor related CLI packages
This commit updates the build scripts, adding a buildPackagesAndLinkCli function for improved CLI usage. This introduces the 'watch' command, which triggers a build on file changes, improving the development experience.
Several changes have been made to the scripts and several packages, including the addition of get-routes and get-package-json files, and the restructuring of build.ts methods for wider usage.
The HttpClientHandler has been added to manage HTTP requests, and the FirebaseHttpClientProvider is now exported to aid Firebase integration. Also, the api-adapter.ts file has been updated to import HttpRequestHandler.
The changes mainly focus on improving build processes, refactoring the codebase, enhancing the CLI, and allowing broader usage of functions and methods.